refactor: move original scripts to Unity/
