+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Unity/Logger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 01adb9089f997797886f579aeae2c08b
+guid: 0ac29e34e770fcf2da57e7f805844165
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
