+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Unity/ResourceManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e12e29c550ae13b03972c9778f7be338
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
