+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Unity/ResourceManager/AssetBundleResourceManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 41e48e5f9be463ba2b2480266808f889
+guid: 39737784a5d929cc4ab94aa5f36635c2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
