+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Unity/ResourceManager/LocalResourceManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a143a93e8ee4f9f386f546e6038566d
+guid: 987096dd33929458589fc4bfd0a199c4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
