+++ Packages/com.github.homuler.mediapipe/Runtime/Scripts/Unity/ResourceManager/StreamingAssetsResourceManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2843176d3bd2776b4a0e5e3038b30831
+guid: 8590959d1f3a5951792fb898918d64d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
